added strategy pattern
